Refactor codebase for better structure and maintainability. relates to #33
