+++ operate/screen_capture.py
@@ -1,0 +1,95 @@
+from PIL import ImageGrab
+import os 
+import platform
+import subprocess
+import pyautogui
+import Xlib.display
+
+from PIL import Image, ImageDraw, ImageFont
+import matplotlib.font_manager as fm
+
+def capture_screen_with_cursor(file_path=os.path.join("screenshots", "screenshot.png")):
+    user_platform = platform.system()
+    
+    if user_platform == "Windows":
+        screenshot = pyautogui.screenshot()
+        screenshot.save(file_path)
+    elif user_platform == "Linux":
+        # Use xlib to prevent scrot dependency for Linux
+        screen = Xlib.display.Display().screen()
+        size = screen.width_in_pixels, screen.height_in_pixels
+        screenshot = ImageGrab.grab(bbox=(0, 0, size[0], size[1]))
+        screenshot.save(file_path)
+    elif user_platform == "Darwin": # (Mac OS)
+        # Use the screencapture utility to capture the screen with the cursor
+        subprocess.run(["screencapture", "-C", file_path])
+    else:
+        print(f"The platform you're using ({user_platform}) is not currently supported")
+
+def add_grid_to_image(original_image_path, new_image_path, grid_interval):
+    """
+    Add a grid to an image
+    """
+    # Load the image
+    image = Image.open(original_image_path)
+
+    # Create a drawing object
+    draw = ImageDraw.Draw(image)
+
+    # Get the image size
+    width, height = image.size
+
+    # Get the path to a TrueType font included with matplotlib
+    font_paths = fm.findSystemFonts(fontpaths=None, fontext="ttf")
+    # Filter for specific font name (e.g., 'Arial.ttf')
+    font_path = next((path for path in font_paths if "Arial" in path), None)
+    if not font_path:
+        if len(font_paths) > 0:
+            font_path = font_paths[0]
+        else:
+            raise RuntimeError("No TrueType fonts found on the system.")
+
+    # Reduce the font size a bit
+    font_size = int(grid_interval / 10)  # Reduced font size
+    font = ImageFont.truetype(font_path, size=font_size)
+
+    # Calculate the background size based on the font size
+    bg_width = int(font_size * 4.2)  # Adjust as necessary
+    bg_height = int(font_size * 1.2)  # Adjust as necessary
+
+    # Function to draw text with a white rectangle background
+    def draw_label_with_background(position, text, draw, font, bg_width, bg_height):
+        # Adjust the position based on the background size
+        text_position = (position[0] + bg_width // 2, position[1] + bg_height // 2)
+        # Draw the text background
+        draw.rectangle(
+            [position[0], position[1], position[0] + bg_width, position[1] + bg_height],
+            fill="white",
+        )
+        # Draw the text
+        draw.text(text_position, text, fill="black", font=font, anchor="mm")
+
+    # Draw vertical lines and labels at every `grid_interval` pixels
+    for x in range(grid_interval, width, grid_interval):
+        line = ((x, 0), (x, height))
+        draw.line(line, fill="blue")
+        for y in range(grid_interval, height, grid_interval):
+            # Calculate the percentage of the width and height
+            x_percent = round((x / width) * 100)
+            y_percent = round((y / height) * 100)
+            draw_label_with_background(
+                (x - bg_width // 2, y - bg_height // 2),
+                f"{x_percent}%,{y_percent}%",
+                draw,
+                font,
+                bg_width,
+                bg_height,
+            )
+
+    # Draw horizontal lines - labels are already added with vertical lines
+    for y in range(grid_interval, height, grid_interval):
+        line = ((0, y), (width, y))
+        draw.line(line, fill="blue")
+
+    # Save the image with the grid
+    image.save(new_image_path)
