+++ operate/styling.py
@@ -1,0 +1,27 @@
+from prompt_toolkit.styles import Style as PromptStyle
+
+# Define style
+style = PromptStyle.from_dict(
+    {
+        "dialog": "bg:#88ff88",
+        "button": "bg:#ffffff #000000",
+        "dialog.body": "bg:#44cc44 #ffffff",
+        "dialog shadow": "bg:#003800",
+    }
+)
+# Standard green text
+ANSI_GREEN = "\033[32m"
+# Bright/bold green text
+ANSI_BRIGHT_GREEN = "\033[92m"
+# Reset to default text color
+ANSI_RESET = "\033[0m"
+# ANSI escape code for blue text
+ANSI_BLUE = "\033[94m"  # This is for bright blue
+
+# Standard yellow text
+ANSI_YELLOW = "\033[33m"
+
+ANSI_RED = "\033[31m"
+
+# Bright magenta text
+ANSI_BRIGHT_MAGENTA = "\033[95m"
