+++ operate/utils.py
@@ -1,0 +1,190 @@
+import json
+import os
+import time 
+import base64
+import re
+
+from operate.exceptions import ModelNotRecognizedException
+from operate.config import setup_client
+from operate.screen_capture import capture_screen_with_cursor, add_grid_to_image
+from operate.prompt_template import format_vision_prompt, format_summary_prompt
+
+client = setup_client()
+
+def convert_percent_to_decimal(percent_str):
+    try:
+        # Remove the '%' sign and convert to float
+        decimal_value = float(percent_str.strip("%"))
+
+        # Convert to decimal (e.g., 20% -> 0.20)
+        return decimal_value / 100
+    except ValueError as e:
+        print(f"Error converting percent to decimal: {e}")
+        return None
+    
+def extract_json_from_string(s):
+    # print("extracting json from string", s)
+    try:
+        # Find the start of the JSON structure
+        json_start = s.find("{")
+        if json_start == -1:
+            return None
+
+        # Extract the JSON part and convert it to a dictionary
+        json_str = s[json_start:]
+        return json.loads(json_str)
+    except Exception as e:
+        print(f"Error parsing JSON: {e}")
+        return None
+    
+def get_next_action(model, messages, objective):
+    if model == "gpt-4-vision-preview":
+        content = get_next_action_from_openai(messages, objective)
+        return content
+    elif model == "agent-1":
+        return "coming soon"
+
+    raise ModelNotRecognizedException(model)
+
+
+def get_last_assistant_message(messages):
+    """
+    Retrieve the last message from the assistant in the messages array.
+    If the last assistant message is the first message in the array, return None.
+    """
+    for index in reversed(range(len(messages))):
+        if messages[index]["role"] == "assistant":
+            if index == 0:  # Check if the assistant message is the first in the array
+                return None
+            else:
+                return messages[index]
+    return None  # Return None if no assistant message is found
+
+
+def get_next_action_from_openai(messages, objective):
+    """
+    Get the next action for Self-Operating Computer
+    """
+    # sleep for a second
+    time.sleep(1)
+    try:
+        screenshots_dir = "screenshots"
+        if not os.path.exists(screenshots_dir):
+            os.makedirs(screenshots_dir)
+
+        screenshot_filename = os.path.join(screenshots_dir, "screenshot.png")
+        # Call the function to capture the screen with the cursor
+        capture_screen_with_cursor(screenshot_filename)
+
+        new_screenshot_filename = os.path.join("screenshots", "screenshot_with_grid.png")
+
+        add_grid_to_image(screenshot_filename, new_screenshot_filename, 500)
+        # sleep for a second
+        time.sleep(1)
+
+        with open(new_screenshot_filename, "rb") as img_file:
+            img_base64 = base64.b64encode(img_file.read()).decode("utf-8")
+
+        previous_action = get_last_assistant_message(messages)
+
+        vision_prompt = format_vision_prompt(objective, previous_action)
+
+        vision_message = {
+            "role": "user",
+            "content": [
+                {"type": "text", "text": vision_prompt},
+                {
+                    "type": "image_url",
+                    "image_url": {"url": f"data:image/jpeg;base64,{img_base64}"},
+                },
+            ],
+        }
+        # create a copy of messages and save to pseudo_messages
+        pseudo_messages = messages.copy()
+        pseudo_messages.append(vision_message)
+
+        response = client.chat.completions.create(
+            model="gpt-4-vision-preview",
+            messages=pseudo_messages,
+            presence_penalty=1,
+            frequency_penalty=1,
+            temperature=0.7,
+            max_tokens=300,
+        )
+
+        messages.append(
+            {
+                "role": "user",
+                "content": "`screenshot.png`",
+            }
+        )
+        content = response.choices[0].message.content
+        return content
+
+    except Exception as e:
+        print(f"Error parsing JSON: {e}")
+        return "Failed take action after looking at the screenshot"
+
+
+def parse_oai_response(response):
+    if response == "DONE":
+        return {"type": "DONE", "data": None}
+    elif response.startswith("CLICK"):
+        # Adjust the regex to match the correct format
+        click_data = re.search(r"CLICK \{ (.+) \}", response).group(1)
+        click_data_json = json.loads(f"{{{click_data}}}")
+        return {"type": "CLICK", "data": click_data_json}
+
+    elif response.startswith("TYPE"):
+        # Extract the text to type
+        type_data = re.search(r'TYPE "(.+)"', response, re.DOTALL).group(1)
+        return {"type": "TYPE", "data": type_data}
+
+    elif response.startswith("SEARCH"):
+        # Extract the search query
+        search_data = re.search(r'SEARCH "(.+)"', response).group(1)
+        return {"type": "SEARCH", "data": search_data}
+
+    return {"type": "UNKNOWN", "data": response}
+
+
+def summarize(messages, objective):
+    try:
+        screenshots_dir = "screenshots"
+        if not os.path.exists(screenshots_dir):
+            os.makedirs(screenshots_dir)
+
+        screenshot_filename = os.path.join(screenshots_dir, "summary_screenshot.png")
+        # Call the function to capture the screen with the cursor
+        capture_screen_with_cursor(screenshot_filename)
+
+        with open(screenshot_filename, "rb") as img_file:
+            img_base64 = base64.b64encode(img_file.read()).decode("utf-8")
+
+        summary_prompt = format_summary_prompt(objective)
+
+        summary_message = {
+            "role": "user",
+            "content": [
+                {"type": "text", "text": summary_prompt},
+                {
+                    "type": "image_url",
+                    "image_url": {"url": f"data:image/jpeg;base64,{img_base64}"},
+                },
+            ],
+        }
+        # create a copy of messages and save to pseudo_messages
+        messages.append(summary_message)
+
+        response = client.chat.completions.create(
+            model="gpt-4-vision-preview",
+            messages=messages,
+            max_tokens=500,
+        )
+
+        content = response.choices[0].message.content
+        return content
+
+    except Exception as e:
+        print(f"Error parsing JSON: {e}")
+        return "Failed to summarize the workflow"
